boardSeq 유효성 체크로직 postService -> boardService 이동
