+++ src/test/java/com/example/board/service/impl/BoardListServiceImplTest.java
@@ -3,6 +3,7 @@ package com.example.board.service.impl;
 import com.example.board.model.Board;
 import com.example.common.dto.PageList;
 import com.example.common.dto.PageRequest;
+import com.example.common.exception.DataNotFoundException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,6 +11,8 @@ import org.springframework.boot.test.context.SpringBootTest;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 
 
 @SpringBootTest
@@ -34,5 +37,13 @@ class BoardListServiceImplTest {
         assertThat(result.getTotalCount(), is(12));
         assertThat(result.getTotalPage(), is(2));
         assertThat(result.getList().get(0).getName(), is("KKK"));
+    }
+
+    @Test
+    @DisplayName("validateBoardSeq :: 예외 케이스")
+    void validateBoardSeq() {
+        DataNotFoundException thrown = assertThrows(DataNotFoundException.class,
+                () -> boardService.validateBoardSeq(14));
+        assertEquals("Data not found", thrown.getMessage());
     }
 }
